requirements.txt
Flask==3.0.3
flask-cors==4.0.0
mysql-connector-python==8.4.0
